Set RLP test suite logger to GinkgoWriter
[#137746159]

+++ src/plumbing/conversion/tov2.go
@@ -156,7 +156,7 @@ func convertContainerMetric(v2e *v2.Envelope, e *events.Envelope) {
 				},
 				"cpu": {
 					Unit:  "percentage",
-					Value: float64(t.GetCpuPercentage()),
+					Value: t.GetCpuPercentage(),
 				},
 				"memory": {
 					Unit:  "bytes",
